requirements.txt

torch
torchvision
roma
evo
gradio
matplotlib
tqdm
opencv-python
scipy
einops
trimesh
tensorboard
pyglet<2
huggingface-hub[torch]>=0.22
gdown
imageio
cython
icecream
numpy
Pillow
diffusers
